requirements.txt
fastapi
requests
pymongo==4.11.3
python-jose==3.4.0